Check in new changes

+++ Huddle.MetricWebApp/Models/MetricValue.cs
@@ -9,7 +9,7 @@ namespace Huddle.MetricWebApp.Models
     {
         public int Id { get; set; }
         public Metric Metric{ get; set; }
-        public int? Value { get; set; }
+        public double? Value { get; set; }
         public DateTime InputDate { get; set; }
     }
 }
